run_program: ensure safe failure when stack is small
In the case that only one element is present and a swap or combine is
attempted, an error will raise on the second pop(), and the contents of the
first pop() will be silently dropped. Rewrite these ops such that they attempt
to put contents back on the stack before returning an error.

Signed-off-by: Nicky Sielicki <sielicki@yandex.com>